chore: move polymorphic to utils
