+++ src/ActionList/LinkItem.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {ForwardRefComponent as PolymorphicForwardRefComponent} from '../internal/polymorphic'
+import {ForwardRefComponent as PolymorphicForwardRefComponent} from '../utils/polymorphic'
 import Link from '../Link'
 import {SxProp, merge} from '../sx'
 import {Item, ActionListItemProps} from './Item'
