+++ src/TextInput.tsx
@@ -1,5 +1,5 @@
 import React, {MouseEventHandler, useCallback, useState} from 'react'
-import {ForwardRefComponent as PolymorphicForwardRefComponent} from './internal/polymorphic'
+import {ForwardRefComponent as PolymorphicForwardRefComponent} from './utils/polymorphic'
 import classnames from 'classnames'
 
 import TextInputInnerVisualSlot from './_TextInputInnerVisualSlot'
